v0.0.14 - refactor resolver type location
