+++ src/data_sources/http/filter/mod.rs
@@ -9,9 +9,7 @@ pub mod request_body;
 pub mod url_path;
 pub mod url_search_query;
 
-use crate::{
-    configuration::subgraph::entities::ServiceEntityConfig, graphql::schema::ResolverType,
-};
+use crate::{configuration::subgraph::entities::ServiceEntityConfig, resolver_type::ResolverType};
 
 use super::HttpDataSource;
