+++ src/data_sources/http/filter/request_body/mod.rs
@@ -2,7 +2,7 @@ use async_graphql::Json;
 use bson::Document;
 use log::debug;
 
-use crate::{data_sources::http::HttpDataSource, graphql::schema::ResolverType};
+use crate::{data_sources::http::HttpDataSource, resolver_type::ResolverType};
 
 impl HttpDataSource {
     pub fn create_body_filters(
