+++ src/data_sources/sql/create_query/create_nested_query_recursive/mod.rs
@@ -7,7 +7,7 @@ use crate::{
         data_sources::sql::DialectEnum, entities::ServiceEntityConfig, SubGraphConfig,
     },
     data_sources::sql::{SqlDataSource, SqlValueEnum},
-    graphql::schema::ResolverType,
+    resolver_type::ResolverType,
 };
 
 use super::JoinClauses;
