+++ src/graphql/input/get_entity_field_type/get_entity_object_id_field_type/mod.rs
@@ -1,6 +1,6 @@
 use async_graphql::dynamic::TypeRef;
 
-use crate::graphql::{input::ServiceInput, schema::ResolverType};
+use crate::{graphql::input::ServiceInput, resolver_type::ResolverType};
 
 impl ServiceInput {
     pub fn get_entity_object_id_field_type(
