+++ src/graphql/resolver/create_resolver_function/get_operation_type/mod.rs
@@ -2,7 +2,7 @@ use log::debug;
 
 use crate::{
     configuration::subgraph::entities::service_entity_field::ServiceEntityFieldConfig,
-    graphql::{resolver::ServiceResolver, schema::ResolverType},
+    graphql::resolver::ServiceResolver, resolver_type::ResolverType,
 };
 
 impl ServiceResolver {
