+++ src/graphql/resolver/get_resolver_input_name/mod.rs
@@ -1,6 +1,6 @@
 use log::debug;
 
-use crate::graphql::schema::ResolverType;
+use crate::resolver_type::ResolverType;
 
 use super::ServiceResolver;
